error handling

+++ backend/gql/src/models/stations/station.resolvers.js
@@ -40,6 +40,7 @@ const stationResolvers = {
             try {
                 if (!context.isAdmin) throw context.AuthenticationError;
                 const station = await Stations.findOne({ where: { slug: args.slug } });
+                if (!station) throw new GraphQLError('Station not found', { extensions: { code: ApolloServerErrorCode.INTERNAL_SERVER_ERROR } });
                 await Slot.destroy({ where: { station_id: station.id } });
                 await station.destroy();
                 return station;
@@ -53,9 +54,9 @@ const stationResolvers = {
             try {
                 if (!context.isAdmin) throw context.AuthenticationError;
                 const { slug, name, status, image, latitude, longitude } = args;
-                if (!name && !status && !image && !latitude && !longitude) throw new GraphQLError('No data provided for station update', { extensions: { code: ApolloServerErrorCode.INTERNAL_SERVER_ERROR } });
+                if (!name && !status && !image && !latitude && !longitude) throw new GraphQLError("No data to update", { extensions: { code: ApolloServerErrorCode.BAD_USER_INPUT } });
                 const station = await Stations.findOne({ where: { slug } });
-                if (!station) throw new Error('Station not found');
+                if (!station) throw new GraphQLError('Station not found', { extensions: { code: ApolloServerErrorCode.INTERNAL_SERVER_ERROR } });
                 if (name) station.name = name;
                 if (status) station.status = status;
                 if (image) station.image = image;
